поменял структуру
